feedback 4. 타입 안전하게 관리 및 너무 긴 mockdata파일 분리
