+++ src/main.tsx
@@ -1,10 +1,9 @@
-// src/main.tsx
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import App from '@/App';
 import { GlobalStyles } from '@/styles/global';
 import { ThemeProvider } from '@emotion/react';
-import { theme } from './styles/theme';
+import { theme } from '@/styles/theme';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
